added RefPath objects
