+++ modules/swagger-core/src/main/java/io/swagger/util/Yaml.java
@@ -9,6 +9,7 @@ import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import io.swagger.models.Model;
+import io.swagger.models.Path;
 import io.swagger.models.auth.SecuritySchemeDefinition;
 import io.swagger.models.parameters.Parameter;
 import io.swagger.models.properties.Property;
@@ -21,6 +22,7 @@ public class Yaml {
             mapper = new ObjectMapper(new YAMLFactory());
             // mapper.registerModule(new JodaModule());
             SimpleModule module = new SimpleModule();
+            module.addDeserializer(Path.class, new PathDeserializer());
             module.addDeserializer(Model.class, new ModelDeserializer());
             module.addDeserializer(Property.class, new PropertyDeserializer());
             module.addDeserializer(Parameter.class, new ParameterDeserializer());
