+++ modules/swagger-core/src/test/scala/deserialization/JsonDesrializationTest.scala
@@ -1,4 +1,4 @@
-import io.swagger.models.Swagger
+import io.swagger.models.{PathImpl, RefPath, Swagger}
 import io.swagger.util.Json
 import org.junit.runner.RunWith
 import org.scalatest.{FlatSpec, Matchers}
@@ -18,5 +18,13 @@ class JsonDesrializationTest extends FlatSpec with Matchers {
     c.isInstanceOf[Swagger] should be(true)
 
     // Json.prettyPrint(c)
+  }
+
+  it should "deserialize a path ref" in {
+    val c = m.readValue(new java.io.File("src/test/scala/specFiles/pathRef.json"), classOf[Swagger])
+    c.isInstanceOf[Swagger] should be(true)
+
+    c.asInstanceOf[Swagger].getPath("/pet").isInstanceOf[RefPath] should be(true)
+    c.asInstanceOf[Swagger].getPath("/user").isInstanceOf[PathImpl] should be(true)
   }
 }
