+++ modules/swagger-models/src/main/java/io/swagger/models/refs/RefType.java
@@ -5,5 +5,5 @@ package io.swagger.models.refs;
  */
 public enum RefType {
     DEFINITION,
-    PARAMETER
+    PATH, PARAMETER
 }
